Add spm support
