+++ Sources/StyledText/Classes/WeakArray.swift
