+++ StyledText.podspec
@@ -11,6 +11,6 @@ Pod::Spec.new do |s|
   s.source           = { :git => 'https://github.com/blueapron/styled-text.git', :tag => 'v' + s.version.to_s }
   s.ios.deployment_target = '10.0'
   s.swift_version = '5.0'
-  s.source_files = 'StyledText/Classes/**/*'
+  s.source_files = 'Sources/StyledText/Classes/**/*'
   s.frameworks = 'UIKit'
 end
